Create com.uic.medialab.sh
Take II lol, its late. Creating this for the Media Lab Loaners, which are going to do a lite-version of the Staff Mac process using DEPNotify so they can function a bit more like a Staff Mac and not a locked down lab machine. Will be wiped during return process.